Plan 9 from Bell Labs 1999-02-09

+++ port/alloc.c
@@ -181,7 +181,7 @@ pooladd(Pool *p, Bhdr *q)
 }
 
 void*
-poolalloc(Pool *p, int size)
+poolalloc(Pool *p, long size)
 {
 	Bhdr *q, *t;
 	int alloc, ldr, ns, frag;
